move all ImageItem classes into its own directory
