+++ libheif/codecs/hevc_boxes.h
@@ -28,7 +28,7 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include <codecs/image_item.h>
+#include "image-items/image_item.h"
 
 
 class Box_hvcC : public Box
