+++ libheif/codecs/vvc_boxes.h
@@ -24,7 +24,7 @@
 #include "box.h"
 #include <string>
 #include <vector>
-#include <codecs/image_item.h>
+#include "image-items/image_item.h"
 #include <memory>
