+++ libheif/image-items/avc.h
@@ -1,6 +1,6 @@
 /*
- * HEIF codec.
- * Copyright (c) 2017 Dirk Farin <dirk.farin@gmail.com>
+ * HEIF AVC codec.
+ * Copyright (c) 2023 Brad Hards <bradh@frogmouth.net>
  *
  * This file is part of libheif.
  *
@@ -18,37 +18,37 @@
  * along with libheif.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef HEIF_AVIF_H
-#define HEIF_AVIF_H
+#ifndef HEIF_AVC_H
+#define HEIF_AVC_H
 
-#include <cassert>
-#include <cmath>
-#include <memory>
-#include <string>
-#include <vector>
-
-#include "libheif/heif.h"
 #include "box.h"
 #include "error.h"
-#include <codecs/image_item.h>
+#include <cstdint>
+#include <vector>
+#include <string>
+#include <memory>
+#include "image_item.h"
 
 
-class ImageItem_AVIF : public ImageItem
+class ImageItem_AVC : public ImageItem
 {
 public:
-  ImageItem_AVIF(HeifContext* ctx, heif_item_id id) : ImageItem(ctx, id) {}
+  ImageItem_AVC(HeifContext* ctx, heif_item_id id) : ImageItem(ctx, id) {}
 
-  ImageItem_AVIF(HeifContext* ctx) : ImageItem(ctx) {}
+  ImageItem_AVC(HeifContext* ctx) : ImageItem(ctx) {}
 
-  uint32_t get_infe_type() const override { return fourcc("av01"); }
+  uint32_t get_infe_type() const override { return fourcc("avc1"); }
 
   const char* get_auxC_alpha_channel_type() const override { return "urn:mpeg:mpegB:cicp:systems:auxiliary:alpha"; }
 
   const heif_color_profile_nclx* get_forced_output_nclx() const override { return nullptr; }
 
-  heif_compression_format get_compression_format() const override { return heif_compression_AV1; }
+  heif_compression_format get_compression_format() const override { return heif_compression_AVC; }
 
   Error on_load_file() override;
+
+protected:
+  std::shared_ptr<Decoder> get_decoder() const override;
 
 public:
   Result<CodedImageData> encode(const std::shared_ptr<HeifPixelImage>& image,
@@ -56,13 +56,7 @@ public:
                                 const struct heif_encoding_options& options,
                                 enum heif_image_input_class input_class) override;
 
-protected:
-  Result<std::vector<uint8_t>> read_bitstream_configuration_data(heif_item_id itemId) const override;
-
-  std::shared_ptr<class Decoder> get_decoder() const override;
-
-private:
-  std::shared_ptr<class Decoder_AVIF> m_decoder;
+  std::shared_ptr<class Decoder_AVC> m_decoder;
 };
 
 #endif
