+++ libheif/image-items/grid.cc
@@ -19,8 +19,8 @@
  */
 
 #include "grid.h"
-#include <context.h>
-#include <file.h>
+#include "context.h"
+#include "file.h"
 #include <cstring>
 #include <deque>
 #include <future>
