+++ libheif/image-items/hevc.h
@@ -28,7 +28,7 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include <codecs/image_item.h>
+#include "image_item.h"
 
 
 class ImageItem_HEVC : public ImageItem
