+++ libheif/image-items/jpeg.h
@@ -24,7 +24,7 @@
 #include "box.h"
 #include <string>
 #include <vector>
-#include <codecs/image_item.h>
+#include "image_item.h"
 #include <memory>
