+++ libheif/image-items/mask_image.h
