+++ libheif/image-items/overlay.cc
