+++ libheif/image-items/tild.h
@@ -22,7 +22,7 @@
 #define LIBHEIF_TILD_H
 
 
-#include <codecs/image_item.h>
+#include "image_item.h"
 #include "box.h"
 #include <vector>
 #include <string>
