+++ libheif/image-items/unc_image.h
