+++ libheif/image-items/vvc.h
@@ -1,6 +1,6 @@
 /*
- * HEIF AVC codec.
- * Copyright (c) 2023 Brad Hards <bradh@frogmouth.net>
+ * HEIF VVC codec.
+ * Copyright (c) 2023 Dirk Farin <dirk.farin@gmail.com>
  *
  * This file is part of libheif.
  *
@@ -18,45 +18,45 @@
  * along with libheif.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef HEIF_AVC_H
-#define HEIF_AVC_H
+#ifndef LIBHEIF_VVC_H
+#define LIBHEIF_VVC_H
 
 #include "box.h"
-#include "error.h"
-#include <cstdint>
-#include <vector>
 #include <string>
+#include <vector>
+#include "image_item.h"
 #include <memory>
-#include "codecs/image_item.h"
 
 
-class ImageItem_AVC : public ImageItem
+class ImageItem_VVC : public ImageItem
 {
 public:
-  ImageItem_AVC(HeifContext* ctx, heif_item_id id) : ImageItem(ctx, id) {}
+  ImageItem_VVC(HeifContext* ctx, heif_item_id id) : ImageItem(ctx, id) {}
 
-  ImageItem_AVC(HeifContext* ctx) : ImageItem(ctx) {}
+  ImageItem_VVC(HeifContext* ctx) : ImageItem(ctx) {}
 
-  uint32_t get_infe_type() const override { return fourcc("avc1"); }
+  uint32_t get_infe_type() const override { return fourcc("vvc1"); }
 
   const char* get_auxC_alpha_channel_type() const override { return "urn:mpeg:mpegB:cicp:systems:auxiliary:alpha"; }
 
   const heif_color_profile_nclx* get_forced_output_nclx() const override { return nullptr; }
 
-  heif_compression_format get_compression_format() const override { return heif_compression_AVC; }
+  heif_compression_format get_compression_format() const override { return heif_compression_VVC; }
+
+  Result<CodedImageData> encode(const std::shared_ptr<HeifPixelImage>& image,
+                                struct heif_encoder* encoder,
+                                const struct heif_encoding_options& options,
+                                enum heif_image_input_class input_class) override;
 
   Error on_load_file() override;
 
 protected:
   std::shared_ptr<Decoder> get_decoder() const override;
 
-public:
-  Result<CodedImageData> encode(const std::shared_ptr<HeifPixelImage>& image,
-                                struct heif_encoder* encoder,
-                                const struct heif_encoding_options& options,
-                                enum heif_image_input_class input_class) override;
+  Result<std::vector<uint8_t>> read_bitstream_configuration_data(heif_item_id itemId) const override;
 
-  std::shared_ptr<class Decoder_AVC> m_decoder;
+private:
+  std::shared_ptr<class Decoder_VVC> m_decoder;
 };
 
-#endif
+#endif // LIBHEIF_VVC_H
